move constants to wutil
